Move PluginMessage API into InventorySetupsPluginMessageHandler
Review feedback on #416:
- Extract the API constants, the volatile name snapshot and all message
  handling out of InventorySetupsPlugin into a dedicated
  InventorySetupsPluginMessageHandler, constructed in startUp like the
  other helpers. The plugin keeps a thin @Subscribe that delegates, and
  a one-line broadcastSetupsChanged for the data manager.
- Split the namespace / own-broadcast guard into two if statements.
- Add a default case that logs unsupported message names at debug.
- Rework the clear name check into explicit branches: a null name
  clears whatever is active; a given name only clears when it matches
  the currently shown setup.

No behavior change; the message contract is identical.
