+++ src/main/java/inventorysetups/InventorySetupsPlugin.java
@@ -181,24 +181,6 @@ public class InventorySetupsPlugin extends Plugin
 	private static final int SPELLBOOK_VARBIT = 4070;
 	private static final int BANK_TAG_OPTIONS = BankTagsService.OPTION_ALLOW_MODIFICATIONS | BankTagsService.OPTION_HIDE_TAG_NAME;
 
-	// PluginMessage API for other plugins to list and open/clear inventory setups. See #415.
-	public static final String API_NAMESPACE = "inventory-setups";
-	// Bumped when the contract below changes in a breaking way. Shipped in setups-changed as data["version"].
-	public static final int API_VERSION = 1;
-	// out: broadcast when the setups change. data["setups"] = List<String> of names, data["version"] = int.
-	public static final String API_MSG_SETUPS_CHANGED = "setups-changed";
-	// in: list setups on demand (for plugins that start after us). Put a mutable Collection<String> under
-	// "setups"; it is filled synchronously with the current setup names.
-	public static final String API_MSG_GET_SETUPS = "get-setups";
-	// in: open a setup, filtering the bank like the worn items menu. data["setup"] = name.
-	public static final String API_MSG_VIEW = "view";
-	// in: clear the current setup (like worn items "Close current setup"). data["setup"] = name to clear
-	// only when it is the active setup; omit to clear whatever is active.
-	public static final String API_MSG_CLEAR = "clear";
-	public static final String API_DATA_SETUPS = "setups";
-	public static final String API_DATA_SETUP = "setup";
-	public static final String API_DATA_VERSION = "version";
-
 	@Inject
 	@Getter
 	private Client client;
@@ -306,6 +288,9 @@ public class InventorySetupsPlugin extends Plugin
 	@Getter
 	private InventorySetupsAmmoHandler ammoHandler;
 
+	@Getter
+	private InventorySetupsPluginMessageHandler pluginMessageHandler;
+
 	// Used to defer highlighting to GameTick
 	private boolean shouldTriggerInventoryHighlightOnGameTick;
 
@@ -380,6 +365,7 @@ public class InventorySetupsPlugin extends Plugin
 		this.sections = new ArrayList<>();
 		this.dataManager = new InventorySetupsPersistentDataManager(this, configManager, cache, gson, inventorySetups, sections);
 		this.ammoHandler = new InventorySetupsAmmoHandler(this, client, itemManager, panel, config);
+		this.pluginMessageHandler = new InventorySetupsPluginMessageHandler(this, clientThread, eventBus, panel);
 		this.layoutUtilities = new InventorySetupLayoutUtilities(itemManager, tagManager, layoutManager, config, client);
 		this.canUseLayouts = canUseLayouts();
 
@@ -2678,107 +2664,17 @@ public class InventorySetupsPlugin extends Plugin
 		// config will already be updated by caller so no need to update it here
 	}
 
-	// Immutable snapshot of the setup names, republished on every change. Lets get-setups answer from any
-	// thread without touching the live list. Only written on the mutating thread (see broadcastSetupsChanged).
-	private volatile List<String> setupNamesSnapshot = List.of();
-
-	private List<String> buildSetupNames()
-	{
-		final List<String> names = new ArrayList<>(inventorySetups.size());
-		for (final InventorySetup setup : inventorySetups)
-		{
-			names.add(setup.getName());
-		}
-		return List.copyOf(names);
-	}
-
-	// Refresh the snapshot and notify listeners. Serialized onto the client thread because setups are
-	// mutated from both the client thread and the Swing EDT. Skips the post when the name list is unchanged,
-	// since updateConfig also fires on slot and note edits.
+	// Called by the data manager whenever setups are persisted; the handler decides whether a broadcast
+	// is actually needed.
 	public void broadcastSetupsChanged()
 	{
-		clientThread.invoke(() ->
-		{
-			final List<String> names = buildSetupNames();
-			if (names.equals(setupNamesSnapshot))
-			{
-				return;
-			}
-			setupNamesSnapshot = names;
-			eventBus.post(new PluginMessage(API_NAMESPACE, API_MSG_SETUPS_CHANGED,
-				Map.of(API_DATA_SETUPS, names, API_DATA_VERSION, API_VERSION)));
-		});
+		pluginMessageHandler.broadcastSetupsChanged();
 	}
 
 	@Subscribe
 	public void onPluginMessage(final PluginMessage message)
 	{
-		if (!API_NAMESPACE.equals(message.getNamespace())
-			|| API_MSG_SETUPS_CHANGED.equals(message.getName()))
-		{
-			// Ignore other namespaces and our own outgoing broadcast.
-			return;
-		}
-
-		switch (message.getName())
-		{
-			case API_MSG_GET_SETUPS:
-			{
-				final Object container = message.getData().get(API_DATA_SETUPS);
-				if (container instanceof Collection)
-				{
-					// eventBus.post is synchronous, so the caller's collection is filled before its own
-					// post() call returns.
-					//noinspection unchecked
-					((Collection<String>) container).addAll(setupNamesSnapshot);
-				}
-				break;
-			}
-			case API_MSG_VIEW:
-			{
-				final Object nameObj = message.getData().get(API_DATA_SETUP);
-				if (!(nameObj instanceof String))
-				{
-					break;
-				}
-				final String targetName = (String) nameObj;
-				// Resolve and apply on the client thread, where inventorySetups is otherwise accessed.
-				clientThread.invoke(() ->
-				{
-					final InventorySetup target = inventorySetups.stream()
-						.filter(setup -> setup.getName().equals(targetName))
-						.findFirst()
-						.orElse(null);
-					if (target == null)
-					{
-						log.debug("Ignoring view request for unknown setup '{}'", targetName);
-						return;
-					}
-					panel.setCurrentInventorySetup(target, true);
-				});
-				break;
-			}
-			case API_MSG_CLEAR:
-			{
-				final Object nameObj = message.getData().get(API_DATA_SETUP);
-				clientThread.invoke(() ->
-				{
-					final InventorySetup current = panel.getCurrentSelectedSetup();
-					if (current == null)
-					{
-						return;
-					}
-					// If a setup name is given, only clear when it is the one currently shown, so a caller
-					// never closes a setup the user switched to themselves.
-					if (nameObj instanceof String && !current.getName().equals(nameObj))
-					{
-						return;
-					}
-					panel.returnToOverviewPanel(false);
-				});
-				break;
-			}
-		}
+		pluginMessageHandler.handleMessage(message);
 	}
 
 }
